\name{pcot2-internal}
\alias{aveExprs}
\alias{plotCor}
\alias{t2.permu}
\alias{wb}
\alias{Cor}
\alias{pco}
\alias{rb}
\alias{t2}
\alias{geneLocator}
\alias{rg}
\alias{t2.unequ}

\title{ Internal PCOT2 functions }
\description{
These functions are intended for internal use only.
}
\author{Sarah Song and Mik Black}

\keyword{internal}

